test(expr): move missing type tests
